Organize files for better readability
