+++ lib/ui/views/nft_creation/layouts/components/import_tab/nft_creation_process_import_tab_image.dart
@@ -1,5 +1,5 @@
 /// SPDX-License-Identifier: AGPL-3.0-or-later
-part of '../nft_creation_process_sheet.dart';
+part of '../../nft_creation_process_sheet.dart';
 
 class NFTCreationProcessImportTabImage extends ConsumerWidget {
   const NFTCreationProcessImportTabImage({super.key});
