+++ lib/ui/views/nft_creation/layouts/components/infos_tab/nft_creation_process_infos_tab.dart
@@ -1,5 +1,5 @@
 /// SPDX-License-Identifier: AGPL-3.0-or-later
-part of '../nft_creation_process_sheet.dart';
+part of '../../nft_creation_process_sheet.dart';
 
 class NFTCreationProcessInfosTab extends ConsumerWidget {
   const NFTCreationProcessInfosTab({
