+++ lib/ui/views/nft_creation/layouts/components/infos_tab/nft_creation_process_infos_tab_textfield_name.dart
@@ -1,33 +1,32 @@
 /// SPDX-License-Identifier: AGPL-3.0-or-later
-part of '../nft_creation_process_sheet.dart';
+part of '../../nft_creation_process_sheet.dart';
 
-class NFTCreationProcessInfosTabTextFieldDescription
-    extends ConsumerStatefulWidget {
-  const NFTCreationProcessInfosTabTextFieldDescription({
+class NFTCreationProcessInfosTabTextFieldName extends ConsumerStatefulWidget {
+  const NFTCreationProcessInfosTabTextFieldName({
     super.key,
   });
 
   @override
-  ConsumerState<NFTCreationProcessInfosTabTextFieldDescription> createState() =>
-      _NFTCreationProcessInfosTabTextFieldDescriptionState();
+  ConsumerState<NFTCreationProcessInfosTabTextFieldName> createState() =>
+      _NFTCreationProcessInfosTabTextFieldNameState();
 }
 
-class _NFTCreationProcessInfosTabTextFieldDescriptionState
-    extends ConsumerState<NFTCreationProcessInfosTabTextFieldDescription> {
-  late FocusNode nftDescriptionFocusNode;
-  late TextEditingController nftDescriptionController;
+class _NFTCreationProcessInfosTabTextFieldNameState
+    extends ConsumerState<NFTCreationProcessInfosTabTextFieldName> {
+  late FocusNode nftNameFocusNode;
+  late TextEditingController nftNameController;
 
   @override
   void initState() {
     super.initState();
-    nftDescriptionFocusNode = FocusNode();
-    nftDescriptionController = TextEditingController();
+    nftNameFocusNode = FocusNode();
+    nftNameController = TextEditingController();
   }
 
   @override
   void dispose() {
-    nftDescriptionFocusNode.dispose();
-    nftDescriptionController.dispose();
+    nftNameFocusNode.dispose();
+    nftNameController.dispose();
     super.dispose();
   }
 
@@ -51,25 +50,27 @@ class _NFTCreationProcessInfosTabTextFieldDescriptionState
         ),
       ),
       (_, nftCreation) {
-        if (nftCreation.description != nftDescriptionController.text) {
-          nftDescriptionController.text = nftCreation.description;
+        if (nftCreation.name != nftNameController.text) {
+          nftNameController.text = nftCreation.name;
         }
       },
     );
 
     return AppTextField(
-      focusNode: nftDescriptionFocusNode,
-      controller: nftDescriptionController,
-      textInputAction: TextInputAction.newline,
-      textAlign: TextAlign.start,
+      key: const Key('nftCreationField'),
+      focusNode: nftNameFocusNode,
+      controller: nftNameController,
       cursorColor: theme.text,
-      labelText: AppLocalizations.of(context)!.nftDescriptionHint,
+      textInputAction: TextInputAction.done,
+      labelText: AppLocalizations.of(context)!.nftNameHint,
       autocorrect: false,
-      keyboardType: TextInputType.multiline,
-      maxLines: 4,
+      keyboardType: TextInputType.text,
       style: theme.textStyleSize16W600Primary,
+      inputFormatters: <LengthLimitingTextInputFormatter>[
+        LengthLimitingTextInputFormatter(40),
+      ],
       onChanged: (text) {
-        nftCreationNotifier.setDescription(text);
+        nftCreationNotifier.setName(text);
       },
       suffixButton: hasQRCode
           ? TextFieldButton(
@@ -102,7 +103,7 @@ class _NFTCreationProcessInfosTabTextFieldDescriptionState
                   );
                   return;
                 } else {
-                  nftDescriptionController.text = scanResult;
+                  nftNameController.text = scanResult;
                 }
               },
             )
