+++ lib/ui/views/nft_creation/layouts/components/preview/nft_creation_process_file_preview.dart
@@ -3,10 +3,10 @@ import 'package:aewallet/application/settings/theme.dart';
 import 'package:aewallet/ui/util/styles.dart';
 import 'package:aewallet/ui/views/nft_creation/bloc/provider.dart';
 import 'package:aewallet/ui/views/nft_creation/bloc/state.dart';
-import 'package:aewallet/ui/views/nft_creation/layouts/components/nft_creation_process_file_preview_aeweb.dart';
-import 'package:aewallet/ui/views/nft_creation/layouts/components/nft_creation_process_file_preview_file.dart';
-import 'package:aewallet/ui/views/nft_creation/layouts/components/nft_creation_process_file_preview_http.dart';
-import 'package:aewallet/ui/views/nft_creation/layouts/components/nft_creation_process_file_preview_ipfs.dart';
+import 'package:aewallet/ui/views/nft_creation/layouts/components/preview/nft_creation_process_file_preview_aeweb.dart';
+import 'package:aewallet/ui/views/nft_creation/layouts/components/preview/nft_creation_process_file_preview_file.dart';
+import 'package:aewallet/ui/views/nft_creation/layouts/components/preview/nft_creation_process_file_preview_http.dart';
+import 'package:aewallet/ui/views/nft_creation/layouts/components/preview/nft_creation_process_file_preview_ipfs.dart';
 import 'package:aewallet/util/mime_util.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
