+++ lib/ui/views/nft_creation/layouts/components/properties_access/nft_creation_process_access_label.dart
