+++ lib/ui/views/nft_creation/layouts/components/properties_access/nft_creation_process_file_access.dart
@@ -2,8 +2,8 @@
 import 'package:aewallet/application/settings/settings.dart';
 import 'package:aewallet/application/settings/theme.dart';
 import 'package:aewallet/ui/views/nft_creation/bloc/provider.dart';
-import 'package:aewallet/ui/views/nft_creation/layouts/components/nft_creation_process_access_label.dart';
-import 'package:aewallet/ui/views/nft_creation/layouts/components/nft_creation_process_file_access_add_button.dart';
+import 'package:aewallet/ui/views/nft_creation/layouts/components/properties_access/nft_creation_process_access_label.dart';
+import 'package:aewallet/ui/views/nft_creation/layouts/components/properties_access/nft_creation_process_file_access_add_button.dart';
 import 'package:aewallet/ui/widgets/components/dialog.dart';
 import 'package:aewallet/ui/widgets/components/item_remove_button.dart';
 import 'package:aewallet/util/get_it_instance.dart';
