+++ lib/ui/views/nft_creation/layouts/components/properties_tab/nft_creation_process_properties_list.dart
@@ -1,5 +1,5 @@
 /// SPDX-License-Identifier: AGPL-3.0-or-later
-part of '../nft_creation_process_sheet.dart';
+part of '../../nft_creation_process_sheet.dart';
 
 class NFTCreationProcessPropertiesList extends ConsumerWidget {
   const NFTCreationProcessPropertiesList({
