+++ lib/ui/views/nft_creation/layouts/components/properties_tab/nft_creation_process_property_access.dart
