+++ lib/ui/views/nft_creation/layouts/nft_creation_process_sheet.dart
@@ -11,15 +11,14 @@ import 'package:aewallet/ui/util/styles.dart';
 import 'package:aewallet/ui/util/ui_util.dart';
 import 'package:aewallet/ui/views/nft_creation/bloc/provider.dart';
 import 'package:aewallet/ui/views/nft_creation/bloc/state.dart';
-import 'package:aewallet/ui/views/nft_creation/layouts/components/card_category_with_text.dart';
 import 'package:aewallet/ui/views/nft_creation/layouts/components/category_template_form.dart';
+import 'package:aewallet/ui/views/nft_creation/layouts/components/import_tab/nft_creation_process_import_tab_aeweb_form.dart';
+import 'package:aewallet/ui/views/nft_creation/layouts/components/import_tab/nft_creation_process_import_tab_http_form.dart';
+import 'package:aewallet/ui/views/nft_creation/layouts/components/import_tab/nft_creation_process_import_tab_ipfs_form.dart';
 import 'package:aewallet/ui/views/nft_creation/layouts/components/nft_creation_confirm_sheet.dart';
 import 'package:aewallet/ui/views/nft_creation/layouts/components/nft_creation_form_sheet.dart';
-import 'package:aewallet/ui/views/nft_creation/layouts/components/nft_creation_process_file_preview.dart';
-import 'package:aewallet/ui/views/nft_creation/layouts/components/nft_creation_process_import_tab_aeweb_form.dart';
-import 'package:aewallet/ui/views/nft_creation/layouts/components/nft_creation_process_import_tab_http_form.dart';
-import 'package:aewallet/ui/views/nft_creation/layouts/components/nft_creation_process_import_tab_ipfs_form.dart';
-import 'package:aewallet/ui/views/nft_creation/layouts/components/nft_creation_process_property_access.dart';
+import 'package:aewallet/ui/views/nft_creation/layouts/components/preview/nft_creation_process_file_preview.dart';
+import 'package:aewallet/ui/views/nft_creation/layouts/components/properties_tab/nft_creation_process_property_access.dart';
 import 'package:aewallet/ui/widgets/components/app_button_tiny.dart';
 import 'package:aewallet/ui/widgets/components/app_text_field.dart';
 import 'package:aewallet/ui/widgets/components/icons.dart';
@@ -42,23 +41,23 @@ import 'package:image_picker/image_picker.dart';
 import 'package:mime_dart/mime_dart.dart';
 import 'package:path/path.dart' as path;
 
-part 'components/nft_creation_process_import_tab.dart';
-part 'components/nft_creation_process_import_tab_aeweb.dart';
-part 'components/nft_creation_process_import_tab_camera.dart';
-part 'components/nft_creation_process_import_tab_file.dart';
-part 'components/nft_creation_process_import_tab_http.dart';
-part 'components/nft_creation_process_import_tab_image.dart';
-part 'components/nft_creation_process_import_tab_ipfs.dart';
-part 'components/nft_creation_process_infos_tab.dart';
-part 'components/nft_creation_process_infos_tab_textfield_description.dart';
-part 'components/nft_creation_process_infos_tab_textfield_name.dart';
-part 'components/nft_creation_process_infos_tab_textfield_symbol.dart';
-part 'components/nft_creation_process_properties_list.dart';
-part 'components/nft_creation_process_properties_tab.dart';
-part 'components/nft_creation_process_properties_tab_textfield_name.dart';
-part 'components/nft_creation_process_properties_tab_textfield_search.dart';
-part 'components/nft_creation_process_properties_tab_textfield_value.dart';
+part 'components/import_tab/nft_creation_process_import_tab.dart';
+part 'components/import_tab/nft_creation_process_import_tab_aeweb.dart';
+part 'components/import_tab/nft_creation_process_import_tab_camera.dart';
+part 'components/import_tab/nft_creation_process_import_tab_file.dart';
+part 'components/import_tab/nft_creation_process_import_tab_http.dart';
+part 'components/import_tab/nft_creation_process_import_tab_image.dart';
+part 'components/import_tab/nft_creation_process_import_tab_ipfs.dart';
+part 'components/infos_tab/nft_creation_process_infos_tab.dart';
+part 'components/infos_tab/nft_creation_process_infos_tab_textfield_description.dart';
+part 'components/infos_tab/nft_creation_process_infos_tab_textfield_name.dart';
+part 'components/infos_tab/nft_creation_process_infos_tab_textfield_symbol.dart';
 part 'components/nft_creation_process_summary_tab.dart';
+part 'components/properties_tab/nft_creation_process_properties_list.dart';
+part 'components/properties_tab/nft_creation_process_properties_tab.dart';
+part 'components/properties_tab/nft_creation_process_properties_tab_textfield_name.dart';
+part 'components/properties_tab/nft_creation_process_properties_tab_textfield_search.dart';
+part 'components/properties_tab/nft_creation_process_properties_tab_textfield_value.dart';
 
 class NftCreationProcessSheet extends ConsumerWidget {
   const NftCreationProcessSheet({
